// layers/image/app/types/cropper.d.ts
export interface Size {
  width: number
  height: number
}
export type ImageSize = Size